fix(ci): python-full-coverage — part_014.py crossed the 1500-line guardrail
Attention-clear terminate tests added for PR #805 pushed
test_pr_monitor_runner_coverage_edges_part_014.py to 1517 lines,
failing test_first_party_code_files_stay_under_line_limit in coverage
shard 8. Move the two terminate_*-clears-attention tests into new
part_039.py unchanged so part_014 is 1407 lines with real headroom.
